refactor components into separate folder
