+++ src/pages/vesting/agora.tsx
@@ -1,229 +1,18 @@
-import AddIcon from '@mui/icons-material/Add';
-import CloseIcon from '@mui/icons-material/Close';
-import DoneIcon from '@mui/icons-material/Done';
-import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
-import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
-import SecurityIcon from '@mui/icons-material/Security';
-import StopIcon from '@mui/icons-material/Stop';
-import UpdateIcon from '@mui/icons-material/Update';
-import { Box, Button, Checkbox, Collapse, Container, Divider, FormControl, IconButton, InputLabel, List, ListItem, ListItemIcon, ListItemText, MenuItem, Paper, Select, Stack, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, ToggleButton, ToggleButtonGroup, Tooltip, Typography } from "@mui/material";
+import { Box, Button, Container, FormControl, InputLabel, MenuItem, Paper, Select, Stack, ToggleButton, ToggleButtonGroup, Typography } from "@mui/material";
 import { skipToken } from "@reduxjs/toolkit/dist/query";
 import { useQuery } from "@tanstack/react-query";
-import JSZip from "jszip";
-import NextLink from "next/link";
-import { useRouter } from "next/router";
-import { FC, Fragment, useEffect, useMemo, useState } from "react";
-import { formatEther } from "viem";
+import { useEffect, useMemo, useState } from "react";
 import { useAccount } from "wagmi";
 import { useExpectedNetwork } from "../../features/network/ExpectedNetworkContext";
-import { mapProjectStateIntoGnosisSafeBatch } from "../../features/redux/endpoints/vestingAgoraEndpoints";
-import { rpcApi } from "../../features/redux/store";
-import Amount from "../../features/token/Amount";
 import ConnectionBoundary from "../../features/transactionBoundary/ConnectionBoundary";
-import { TransactionBoundary } from "../../features/transactionBoundary/TransactionBoundary";
-import { TransactionButton, transactionButtonDefaultProps } from "../../features/transactionBoundary/TransactionButton";
-import { TransactionDialogActions, TransactionDialogButton } from "../../features/transactionBoundary/TransactionDialog";
 import { BigLoader } from "../../features/vesting/BigLoader";
-import VestingRow from "../../features/vesting/VestingRow";
 import { useVisibleAddress } from "../../features/wallet/VisibleAddressContext";
 import { useTokenQuery } from "../../hooks/useTokenQuery";
-import { TxBuilder } from "../../libs/gnosis-tx-builder";
 import { NextPageWithLayout } from "../_app";
-import { AgoraResponseData, AllowanceActions, ProjectActions, ProjectsOverview, ProjectState, RoundType, roundTypes } from "../api/agora";
-
-// Updated ActionsList component as a MUI table with checkboxes
-const ActionsList: FC<{
-  actions: (ProjectActions | AllowanceActions)[],
-  tokenSymbol: string | undefined,
-  onSelectionChange?: (selectedActions: (ProjectActions | AllowanceActions)[]) => void }> = ({ actions, tokenSymbol = "", onSelectionChange }
-) => {
-    const [selected, setSelected] = useState<number[]>([]);
-
-    // Update selected actions callback whenever selection changes
-    useEffect(() => {
-        if (onSelectionChange) {
-            const selectedActions = selected.map(index => actions[index]);
-            onSelectionChange(selectedActions);
-        }
-    }, [selected, actions, onSelectionChange]);
-
-    const handleSelectAllClick = (event: React.ChangeEvent<HTMLInputElement>) => {
-        if (event.target.checked) {
-            const newSelected = actions.map((_, index) => index);
-            setSelected(newSelected);
-            return;
-        }
-        setSelected([]);
-    };
-
-    const handleClick = (_event: React.MouseEvent<unknown>, index: number) => {
-        const selectedIndex = selected.indexOf(index);
-        let newSelected: number[] = [];
-
-        if (selectedIndex === -1) {
-            newSelected = newSelected.concat(selected, index);
-        } else if (selectedIndex === 0) {
-            newSelected = newSelected.concat(selected.slice(1));
-        } else if (selectedIndex === selected.length - 1) {
-            newSelected = newSelected.concat(selected.slice(0, -1));
-        } else if (selectedIndex > 0) {
-            newSelected = newSelected.concat(
-                selected.slice(0, selectedIndex),
-                selected.slice(selectedIndex + 1),
-            );
-        }
-
-        setSelected(newSelected);
-    };
-
-    const isSelected = (index: number) => selected.indexOf(index) !== -1;
-
-    // Helper function for formatting amounts
-    const formatAmount = (amount: string) => {
-        const amountBigInt = BigInt(amount);
-        return `${formatEther(amountBigInt)} ${tokenSymbol}`;
-    };
-
-    // Helper function for formatting receiver addresses
-    const formatReceiver = (address: string) => {
-        return `${address.slice(0, 6)}...${address.slice(-4)}`;
-    };
-
-    // Get action details based on action type
-    const getActionDetails = (action: ProjectActions | AllowanceActions) => {
-        let actionType = "";
-        let receiver = "";
-        let amount = "";
-        let fromDate: Date | undefined;
-        let toDate: Date | undefined;
-
-        switch (action.type) {
-            case "create-vesting-schedule":
-                actionType = "Create Vesting Schedule";
-                receiver = formatReceiver(action.payload.receiver);
-                amount = formatAmount(action.payload.totalAmount);
-                fromDate = new Date(action.payload.startDate * 1000);
-                toDate = new Date((action.payload.startDate + action.payload.totalDuration) * 1000);
-                break;
-
-            case "update-vesting-schedule":
-                const prevAmount = formatAmount(action.payload.previousTotalAmount);
-                const newAmount = formatAmount(action.payload.totalAmount);
-                const isDifference = action.payload.previousTotalAmount !== action.payload.totalAmount;
-
-                actionType = "Update Vesting Schedule";
-                receiver = formatReceiver(action.payload.receiver);
-                amount = isDifference ? `${prevAmount} → ${newAmount}` : `${newAmount} (unchanged)`;
-                break;
-
-            case "stop-vesting-schedule":
-                actionType = "Stop Vesting Schedule";
-                receiver = formatReceiver(action.payload.receiver);
-                break;
-
-            case "increase-token-allowance":
-                actionType = "Increase Token Allowance";
-                receiver = formatReceiver(action.payload.receiver);
-                amount = formatAmount(action.payload.allowanceDelta);
-                break;
-
-            case "increase-flow-operator-permissions":
-                actionType = "Increase Flow Operator Permissions";
-                receiver = formatReceiver(action.payload.receiver);
-                amount = `${formatAmount(action.payload.flowRateAllowanceDelta)} per second`;
-                break;
-
-            default:
-                actionType = `Unknown Action: ${(action as any).type}`;
-        }
-
-        return { actionType, receiver, amount, fromDate, toDate };
-    };
-
-    if (actions.length === 0) {
-        return <Typography variant="body2" color="text.secondary">No actions needed</Typography>;
-    }
-
-    return (
-        <TableContainer>
-            <Table size="small" aria-label="actions table">
-                <TableHead>
-                    <TableRow>
-                        <TableCell padding="checkbox">
-                            <Checkbox
-                                indeterminate={selected.length > 0 && selected.length < actions.length}
-                                checked={actions.length > 0 && selected.length === actions.length}
-                                onChange={handleSelectAllClick}
-                                inputProps={{ 'aria-label': 'select all actions' }}
-                            />
-                        </TableCell>
-                        <TableCell>Action Type</TableCell>
-                        <TableCell>Receiver</TableCell>
-                        <TableCell>Amount</TableCell>
-                        <TableCell>From Date</TableCell>
-                        <TableCell>To Date</TableCell>
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {actions.map((action, index) => {
-                        const isItemSelected = isSelected(index);
-                        const { actionType, receiver, amount, fromDate, toDate } = getActionDetails(action);
-
-                        return (
-                            <TableRow
-                                hover
-                                onClick={(event) => handleClick(event, index)}
-                                role="checkbox"
-                                aria-checked={isItemSelected}
-                                tabIndex={-1}
-                                key={index}
-                                selected={isItemSelected}
-                            >
-                                <TableCell padding="checkbox">
-                                    <Checkbox
-                                        checked={isItemSelected}
-                                        inputProps={{ 'aria-labelledby': `action-${index}` }}
-                                    />
-                                </TableCell>
-                                <TableCell id={`action-${index}`}>
-                                    {actionType}
-                                </TableCell>
-                                <TableCell>{receiver}</TableCell>
-                                <TableCell>{amount}</TableCell>
-                                <TableCell>
-                                    {fromDate ? (
-                                        <Tooltip title={
-                                            <>
-                                                {Intl.DateTimeFormat().resolvedOptions().timeZone}
-                                                <br />
-                                                {fromDate.toLocaleString()}
-                                            </>
-                                        } arrow>
-                                            <span>{fromDate.toLocaleDateString()}</span>
-                                        </Tooltip>
-                                    ) : null}
-                                </TableCell>
-                                <TableCell>
-                                    {toDate ? (
-                                        <Tooltip title={
-                                            <>
-                                                {Intl.DateTimeFormat().resolvedOptions().timeZone}
-                                                <br />
-                                                {toDate.toLocaleString()}
-                                            </>
-                                        } arrow>
-                                            <span>{toDate.toLocaleDateString()}</span>
-                                        </Tooltip>
-                                    ) : null}
-                                </TableCell>
-                            </TableRow>
-                        );
-                    })}
-                </TableBody>
-            </Table>
-        </TableContainer>
-    );
-};
+import { AgoraResponseData, AllowanceActions, ProjectActions, RoundType, roundTypes } from "../api/agora";
+import { DownloadGnosisSafeTransactionButton, ExecuteTranchUpdateTransactionButton } from '../../features/vesting/agora/buttons';
+import { ActionsList } from '../../features/vesting/agora/ActionsList';
+import { ProjectsTable } from '../../features/vesting/agora/ProjectsTable';
 
 const AgoraPage: NextPageWithLayout = () => {
     const { visibleAddress } = useVisibleAddress();
@@ -355,7 +144,6 @@ const AgoraPage: NextPageWithLayout = () => {
         return null;
     }
 
-    const currentTranchNo = projectsOverview.tranchPlan.currentTranchCount;
     const areButtonsDisabled = allActions.length === 0;
 
     return (
@@ -407,36 +195,7 @@ const AgoraPage: NextPageWithLayout = () => {
                 Projects Overview
             </Typography>
 
-            <TableContainer component={Paper}>
-                <Table aria-label="collapsible table">
-                    <TableHead>
-                        <TableRow>
-                            <TableCell align="center" colSpan={3}>
-                            </TableCell>
-                            <TableCell align="center" colSpan={6}>
-                                Tranches
-                            </TableCell>
-                        </TableRow>
-                        <TableRow>
-                            <TableCell size="small">
-                            </TableCell>
-                            <TableCell size="small" align="center" sx={{ px: 1.5 }}>Status</TableCell>
-                            <TableCell>Project Name(s)</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 1 ? 'action.hover' : "" }} >Tranch 1</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 2 ? 'action.hover' : "" }} >Tranch 2</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 3 ? 'action.hover' : "" }} >Tranch 3</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 4 ? 'action.hover' : "" }} >Tranch 4</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 5 ? 'action.hover' : "" }} >Tranch 5</TableCell>
-                            <TableCell sx={{ ...tranchColumnSxProps, bgcolor: currentTranchNo === 6 ? 'action.hover' : "" }} >Tranch 6</TableCell>
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {rows.map((row) => (
-                            <Row key={row.agoraEntry.id} currentTranchNo={currentTranchNo} chainId={projectsOverview.chainId} superTokenAddress={projectsOverview.superTokenAddress} state={row} />
-                        ))}
-                    </TableBody>
-                </Table>
-            </TableContainer>
+            <ProjectsTable projectsOverview={projectsOverview} rows={rows} />
 
             <Typography variant="h6" gutterBottom sx={{ mt: 3 }}>
                 Pending Actions ({allActions.length})
@@ -462,394 +221,4 @@ const AgoraPage: NextPageWithLayout = () => {
     );
 };
 
-// # Row
-
-const tranchColumnSxProps = {
-  whiteSpace: 'nowrap',
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-  px: 3,
-  borderRight: 1,
-  borderLeft: 1,
-  borderColor: 'divider'
-} as const;
-
-const TranchCell: FC<{
-  currentTranchNo: number
-  tranchNo: number
-  allocation?: { amount: string }
-  tokenSymbol?: string
-}> = ({ currentTranchNo, tranchNo, allocation, tokenSymbol }) => {
-  return (
-    <TableCell
-      sx={{
-        ...tranchColumnSxProps,
-        bgcolor: currentTranchNo === tranchNo ? 'action.hover' : ""
-      }}
-    >
-      {allocation?.amount ? (
-        <Stack
-          direction="row"
-          justifyContent="space-between"
-          alignItems="center"
-          gap={1.5}
-        >
-          <Amount wei={allocation.amount} disableRounding mono />
-          <Typography variant="tooltip" color="text.secondary">
-            {tokenSymbol}
-          </Typography>
-        </Stack>
-      ) : '—'}
-    </TableCell>
-  );
-};
-
-function Status(props: { state: ProjectState }) {
-
-  const getStatusInfo = () => {
-    if (!props.state.agoraEntry.KYCStatusCompleted) {
-      return { color: "text.disabled", message: "KYC not done" };
-    }
-    else if (props.state.todo.length > 0) {
-      return { color: "warning.main", message: "Actions pending" };
-    } else if (props.state.agoraTotalAmount !== props.state.subgraphTotalAmount) {
-      return { color: "error.main", message: "Mismatch of amounts" };
-    } else {
-      return { color: "success.main", message: "All good" };
-    }
-  };
-
-  const statusInfo = getStatusInfo();
-
-  return (
-    <Tooltip title={statusInfo.message} arrow>
-      <Box
-        sx={{
-          width: 14,
-          height: 14,
-          borderRadius: '50%',
-          bgcolor: statusInfo.color,
-          display: 'inline-block',
-          verticalAlign: 'middle'
-        }}
-      />
-    </Tooltip>
-  );
-}
-
-function Row(props: { currentTranchNo: number, chainId: number, superTokenAddress: string, state: ProjectState }) {
-    const { state, currentTranchNo } = props;
-    const [open, setOpen] = useState(false);
-
-    const allocations = useMemo(() => {
-        return state.agoraEntry.amounts.map((amount, index) => ({
-            tranch: index + 1,
-            amount: amount
-        }));
-    }, [state.agoraEntry.amounts]);
-
-    const { data: token } = useTokenQuery({
-        chainId: props.chainId,
-        id: props.superTokenAddress
-    });
-
-    return (
-        <>
-          <TableRow
-              sx={{
-                  '& > *': { borderBottom: 'unset' },
-                  opacity: state.agoraEntry.KYCStatusCompleted ? 1 : 0.6
-              }}
-          >
-                <TableCell size="small" align="center" sx={{
-                  py: 0.5,
-                  pr: 0,
-                  pl: 1
-                }}>
-                    <IconButton
-                        aria-label="expand row"
-                        size="small"
-                        onClick={() => setOpen(!open)}
-                        sx={{
-                          m: 0,
-                          minWidth: "75px"
-                        }}
-                    >
-                      {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-                    </IconButton>
-                </TableCell>
-                <TableCell size="small" align="center" sx={{ px: 1.5 }}>
-                  <Status state={state} />
-                </TableCell>
-                <TableCell sx={{
-                  whiteSpace: 'nowrap',
-                  maxWidth: '200px',
-                  overflow: 'hidden',
-                  textOverflow: 'ellipsis'
-                }}>
-                  {state.agoraEntry.projectNames.map((name) => (
-                    <Typography
-                      key={name}
-                      component="span"
-                      variant="body2"
-                      noWrap
-                      display="block"
-                    >
-                      {name}
-                    </Typography>
-                  ))}
-                </TableCell>
-                {[1, 2, 3, 4, 5, 6].map(tranchNo => (
-                  <TranchCell
-                    key={tranchNo}
-                    currentTranchNo={currentTranchNo}
-                    tranchNo={tranchNo}
-                    allocation={allocations[tranchNo - 1]}
-                    tokenSymbol={token?.symbol}
-                  />
-                ))}
-            </TableRow>
-
-            <TableRow>
-                <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={9}>
-                    <Collapse in={open} timeout="auto" unmountOnExit>
-                        <Box sx={{ p: 2 }}>
-                            <Box sx={{ ml: 1, mt: 1 }}>
-                                <Box sx={{ display: 'flex', mb: 1 }}>
-                                    <Typography variant="body2" color="text.secondary" sx={{ width: 120 }}>
-                                        Project name(s):
-                                    </Typography>
-                                    <Typography
-                                        variant="body2"
-                                    >
-                                        {state.agoraEntry.projectNames.join(', ')}
-                                    </Typography>
-                                </Box>
-
-                                <Box sx={{ display: 'flex', mb: 1 }}>
-                                    <Typography variant="body2" color="text.secondary" sx={{ width: 120 }}>
-                                        Current wallet:
-                                    </Typography>
-                                    <Typography
-                                        variant="body2mono"
-                                    >
-                                        {state.currentWallet}
-                                    </Typography>
-                                </Box>
-
-                                {state.agoraEntry.wallets.length > 1 && (
-                                    <Box sx={{ display: 'flex' }}>
-                                        <Typography variant="body2" color="text.secondary" sx={{ width: 120 }}>
-                                            Previous wallets:
-                                        </Typography>
-                                        <Box>
-                                            {state.agoraEntry.wallets.slice(0, -1).map((wallet, index) => (
-                                                <Typography
-                                                    key={index}
-                                                    variant="body2"
-                                                    sx={{
-                                                        fontFamily: 'monospace',
-                                                        mb: 0.5
-                                                    }}
-                                                >
-                                                    {wallet}
-                                                </Typography>
-                                            ))}
-                                        </Box>
-                                    </Box>
-                                )}
-
-                                <Box sx={{ display: 'flex', mb: 1 }}>
-                                    <Typography variant="body2" color="text.secondary" sx={{ width: 120 }}>
-                                        Agora total:
-                                    </Typography>
-                                    <Typography variant="body2">
-                                        {state.agoraTotalAmount ? `${formatEther(BigInt(state.agoraTotalAmount))} ${token?.symbol}` : '—'}
-                                    </Typography>
-                                </Box>
-                                <Box sx={{ display: 'flex', mb: 1 }}>
-                                    <Typography variant="body2" color="text.secondary" sx={{ width: 120 }}>
-                                        Vesting total:
-                                    </Typography>
-                                    <Typography variant="body2">
-                                        {state.subgraphTotalAmount ? `${formatEther(BigInt(state.subgraphTotalAmount))} ${token?.symbol}` : '—'}
-                                    </Typography>
-                                </Box>
-                            </Box>
-
-                            {
-                              state.allRelevantSchedules.length > 0 && (
-                                <Box sx={{ mt: 2 }}>
-                                    <Typography variant="body2" color="text.secondary" gutterBottom>
-                                        Vesting schedules:
-                                    </Typography>
-                                    <ProjectVestingSchedulesTables project={state} />
-                                </Box>
-                              )
-                            }
-
-                            {state.todo.length > 0 && (
-                                <Box sx={{ mt: 3 }}>
-                                    <Typography variant="body2" color="text.secondary" gutterBottom>
-                                        Pending Actions ({state.todo.length}):
-                                    </Typography>
-                                    <Paper elevation={1} sx={{ p: 2 }}>
-                                        <ActionsList actions={state.todo} tokenSymbol={token?.symbol} />
-                                    </Paper>
-                                </Box>
-                            )}
-                        </Box>
-                    </Collapse>
-                </TableCell>
-            </TableRow>
-
-        </>
-    );
-}
-
 export default AgoraPage;
-
-type Props = {
-    isDisabled: boolean;
-    projectsOverview: ProjectsOverview;
-    actionsToExecute: (AllowanceActions | ProjectActions)[]
-}
-
-export const DownloadGnosisSafeTransactionButton: FC<Props> = ({
-    isDisabled,
-    projectsOverview,
-    actionsToExecute
-}) => {
-    return (
-        <Button {...transactionButtonDefaultProps} disabled={isDisabled} variant="outlined" onClick={async () => {
-            const zip = new JSZip();
-            const zipName = `execute-tranch-${projectsOverview.tranchPlan.currentTranchCount}_for_safe_tx-builder`;
-            const batchFolder = zip.folder(zipName);
-
-            const transactions = mapProjectStateIntoGnosisSafeBatch(projectsOverview, actionsToExecute);
-
-            const safeTxBuilderJSON = TxBuilder.batch(undefined, transactions, {
-                chainId: projectsOverview.chainId
-            });
-
-            const blob = new Blob([JSON.stringify(safeTxBuilderJSON)], {
-                type: "application/json",
-            });
-
-            batchFolder?.file(`batch.json`, blob);
-
-            const objectURL = URL.createObjectURL(
-                (await batchFolder?.generateAsync({ type: "blob" })) as Blob
-            );
-
-            const a = document.createElement('a');
-            a.href = objectURL;
-            a.download = zipName + ".zip";
-            document.body.appendChild(a);
-            a.click();
-
-            // Clean up by revoking the object URL and removing the link
-            URL.revokeObjectURL(objectURL);
-            document.body.removeChild(a);
-        }}>
-            Download Safe Transaction Builder JSON
-        </Button>
-    )
-}
-
-export const ExecuteTranchUpdateTransactionButton: FC<Props> = ({
-    isDisabled,
-    projectsOverview,
-    actionsToExecute
-}) => {
-    const [executeTranchUpdate, executeTranchUpdateResult] =
-        rpcApi.useExecuteTranchUpdateMutation();
-
-    const isVisible = !executeTranchUpdateResult.isSuccess;
-
-    return (
-        <TransactionBoundary mutationResult={executeTranchUpdateResult}>
-            {({
-                network,
-                getOverrides,
-                setDialogLoadingInfo,
-                setDialogSuccessActions,
-                txAnalytics,
-                closeDialog
-            }) =>
-                isVisible && (
-                    <TransactionButton
-                        dataCy={"create-schedule-tx-button"}
-                        disabled={isDisabled}
-                        onClick={async (signer) => {
-                            setDialogLoadingInfo(
-                                <Typography
-                                    variant="h5"
-                                    color="text.secondary"
-                                    translate="yes"
-                                >
-                                    You are executing a tranch update with {actionsToExecute.length} actions.
-                                </Typography>
-                            );
-
-                            executeTranchUpdate({
-                                signer,
-                                superTokenAddress: projectsOverview.superTokenAddress,
-                                chainId: projectsOverview.chainId,
-                                projectsOverview,
-                                actionsToExecute
-                            }).unwrap();
-
-                            setDialogSuccessActions(
-                                <TransactionDialogActions>
-                                    <TransactionDialogButton
-                                        data-cy="ok-button"
-                                        color="primary"
-                                        onClick={closeDialog}
-                                    >
-                                        OK
-                                    </TransactionDialogButton>
-                                </TransactionDialogActions>
-                            );
-                        }}>
-                        Execute Tranch Update
-                    </TransactionButton>
-                )
-            }
-        </TransactionBoundary>
-    );
-};
-
-const ProjectVestingSchedulesTables: FC<{
-    project: ProjectState
-}> = ({ project }) => {
-    const { network } = useExpectedNetwork();
-
-    const router = useRouter();
-    const openDetails = (id: string) => () =>
-        router.push(`/vesting/${network.slugName}/${id}`);
-
-    if (project.allRelevantSchedules.length === 0) {
-        return null;
-    }
-
-    return (
-        <TableContainer component={Paper}>
-            <Table>
-                <TableBody>
-                    {
-                        project.allRelevantSchedules
-                            .map((vestingSchedule) => (
-                                <VestingRow
-                                    key={vestingSchedule.id}
-                                    network={network}
-                                    vestingSchedule={vestingSchedule}
-                                    onClick={openDetails(vestingSchedule.id)}
-                                />
-                            ))
-                    }
-                </TableBody>
-            </Table>
-        </TableContainer>
-    );
-};
